requirements.txt
deepctr==0.4.1
joblib==0.14.1
numpy==1.15.4
pandas==0.20.3
requests==2.22.0
scikit-learn==0.20.3
tensorflow-gpu==1.4.0
tqdm==4.58.0